Create math modules, starting with Vector3 module
We create a folder math to put all math (or calculate) stuffs
that involve math.
The first module included is a Vector3 extension of THREE.Vector3
module. This extension allow us to use all the math methods already
write in THREE. Add methods to evaluate the volumeSign (and coplanar)
property of vectors.