Create Doubly-Connected Edge List (DCEL)
The main data-structure of this project is DCEL, also known as
HalfEdge Mesh (HEMesh) and is used to holds geometric and
topological information to represent meshes in 3D.

Have three registers: vertices, halfEdges and faces.
Each edge is split in two halfEdges: one from origin to end, and
another from end to origin.
See "Computational Geometry: Algorithms and Applications", Mark
de Berg, Chapter 2.

In order to make easy include (and render) the meshes in a THREE.Scene,
our implementation of DCEL is based on THREE.Vector3, THREE.Face3 and
THREE.Geometry modules.