Update DCEL access and search for vertices
Before we have a binary search for vertices, but doesnt work
because we also have to update the indices in the faces and
in the halfEdges.
So we rewrite the search to O(n) search in the vertices array
and we put an index var into the vertex.

Also the halfEdge, now store the object for origin and end
instead the index.