The render task to 2D and 3D is managed by a single controller.
We decide to use the same three.js scene to render 2D content (in
the plane XY using a orthogonal camera), and the 3D content with
the clasic camera.
This cause that a single controller could render all the stuff
of the project. So, the controller is called a WorkspaceViewController
since the area that we render and show the ouput can be seen as
a worskpace.

In order to that we delete the unused View2D and View3D controllers.